created new testing suite, refactoring
