+++ src/test/java/com/playground/playground/feature/FeatureTest.java
@@ -1,4 +1,4 @@
-package com.playground.playground.data;
+package com.playground.playground.feature;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -23,26 +23,6 @@ public class FeatureTest {
    * This tests the SquareFeatureApplier to ensure the correct points are being squared and appended
    * correctly
    */
-  @Test
-  public void testSquareFeatureApplier() {
-    ArrayList<Object> dataPoint1 =
-        new ArrayList<>(Arrays.asList(new ArrayList<>(Arrays.asList(2.0, 3.0)), 0));
-    ArrayList<Object> dataPoint2 =
-        new ArrayList<>(Arrays.asList(new ArrayList<>(Arrays.asList(4.0, 5.0)), 1));
-    ArrayList<ArrayList<Object>> dataset = new ArrayList<>(Arrays.asList(dataPoint1, dataPoint2));
-
-    SquareFeatureApplier squareFeatureApplier = new SquareFeatureApplier(0);
-    ArrayList<ArrayList<Object>> result = squareFeatureApplier.applyFeature(dataset);
-
-    assertNotNull(result);
-    assertEquals(2, result.size());
-
-    ArrayList<Double> data1 = (ArrayList<Double>) result.get(0).get(0);
-    assertEquals(4.0, data1.get(2));
-
-    ArrayList<Double> data2 = (ArrayList<Double>) result.get(1).get(0);
-    assertEquals(16.0, data2.get(2));
-  }
 
   /** checks that the feature applier is working properly */
   @Test
